# Make sure project generation is deterministic across runs