--- a/Sources/TuistCore/Graph/CircularDependencyLinter.swift
+++ b/Sources/TuistCore/Graph/CircularDependencyLinter.swift
@@ -40,7 +40,7 @@ public class CircularDependencyLinter: CircularDependencyLinting {
         }
         cache.add(project: project)
 
-        for target in project.targets.values.sorted() {
+        for target in project.targets.sorted() {
             try lintTarget(
                 path: path,
                 name: target.name,