--- a/Sources/TuistCore/Graph/ModelExtensions/Project+Core.swift
+++ b/Sources/TuistCore/Graph/ModelExtensions/Project+Core.swift
@@ -10,7 +10,7 @@ extension Project {
     /// - Parameter graph: Dependencies graph.
     /// - Returns: Sorted targets.
     public func sortedTargetsForProjectScheme(graph: Graph) -> [Target] {
-        targets.values.sorted { first, second -> Bool in
+        targets.sorted { first, second -> Bool in
             // First criteria: Test bundles at the end
             if first.product.testsBundle, !second.product.testsBundle {
                 return false