--- a/Sources/TuistDependencies/Mappers/ExternalProjectsPlatformNarrowerGraphMapper.swift
+++ b/Sources/TuistDependencies/Mappers/ExternalProjectsPlatformNarrowerGraphMapper.swift
@@ -25,13 +25,12 @@ public struct ExternalProjectsPlatformNarrowerGraphMapper: GraphMapping { // swi
 
         graph.projects = Dictionary(uniqueKeysWithValues: graph.projects.map { projectPath, project in
             var project = project
-            project.targets = Dictionary(uniqueKeysWithValues: project.targets.map { _, target in
-                let mappedTarget = mapTarget(
+            project.targets = project.targets.map({ target in
+                return mapTarget(
                     target: target,
                     project: project,
                     externalTargetSupportedPlatforms: externalTargetSupportedPlatforms
                 )
-                return (mappedTarget.name, mappedTarget)
             })
             return (projectPath, project)
         })