--- a/Sources/TuistGenerator/Generator/InfoPlistContentProvider.swift
+++ b/Sources/TuistGenerator/Generator/InfoPlistContentProvider.swift
@@ -242,7 +242,7 @@ final class InfoPlistContentProvider: InfoPlistContentProviding {
     }
 
     private func hostTarget(for target: Target, in project: Project) -> Target? {
-        project.targets.values.first {
+        project.targets.first {
             $0.dependencies.contains(where: { dependency in
                 if case let .target(name, _) = dependency, name == target.name {
                     return true