--- a/Sources/TuistGenerator/Generator/ProjectGroups.swift
+++ b/Sources/TuistGenerator/Generator/ProjectGroups.swift
@@ -143,7 +143,7 @@ class ProjectGroups {
     }
 
     private static func extractProjectGroupNames(from project: Project) -> [String] {
-        let groups = [project.filesGroup] + project.targets.values.map(\.filesGroup)
+        let groups = [project.filesGroup] + project.targets.map(\.filesGroup)
         let groupNames: [String] = groups.compactMap {
             switch $0 {
             case let .group(name: groupName):