--- a/Sources/TuistGenerator/Linter/ProjectLinter.swift
+++ b/Sources/TuistGenerator/Linter/ProjectLinter.swift
@@ -49,7 +49,7 @@ class ProjectLinter: ProjectLinting {
 
     private func lintTargets(project: Project) -> [LintingIssue] {
         var issues: [LintingIssue] = []
-        issues.append(contentsOf: project.targets.values.flatMap(targetLinter.lint))
+        issues.append(contentsOf: project.targets.flatMap(targetLinter.lint))
         return issues
     }
 }