--- a/Sources/TuistGenerator/Mappers/ExplicitDependencyGraphMapper.swift
+++ b/Sources/TuistGenerator/Mappers/ExplicitDependencyGraphMapper.swift
@@ -23,7 +23,7 @@ public struct ExplicitDependencyGraphMapper: GraphMapping {
 
         graph.projects = Dictionary(uniqueKeysWithValues: graph.projects.map { projectPath, project in
             var project = project
-            project.targets = project.targets.mapValues { target in
+            project.targets = project.targets.map { target in
                 let graphTarget = GraphTarget(path: projectPath, target: target, project: project)
                 let projectDebugConfigurations = project.settings.configurations.keys
                     .filter { $0.variant == .debug }