--- a/Sources/TuistGenerator/Mappers/GenerateEntitlementsProjectMapper.swift
+++ b/Sources/TuistGenerator/Mappers/GenerateEntitlementsProjectMapper.swift
@@ -23,10 +23,10 @@ public final class GenerateEntitlementsProjectMapper: ProjectMapping {
     public func map(project: Project) throws -> (Project, [SideEffectDescriptor]) {
         logger.debug("Transforming project \(project.name): Synthesizing entitlement files'")
 
-        let results = try project.targets.values
-            .reduce(into: (targets: [String: Target](), sideEffects: [SideEffectDescriptor]())) { results, target in
+        let results = try project.targets
+            .reduce(into: (targets: [Target](), sideEffects: [SideEffectDescriptor]())) { results, target in
                 let (updatedTarget, sideEffects) = try map(target: target, project: project)
-                results.targets[updatedTarget.name] = updatedTarget
+                results.targets.append(updatedTarget)
                 results.sideEffects.append(contentsOf: sideEffects)
             }
         var project = project