--- a/Sources/TuistGenerator/Mappers/GenerateInfoPlistProjectMapper.swift
+++ b/Sources/TuistGenerator/Mappers/GenerateInfoPlistProjectMapper.swift
@@ -37,10 +37,10 @@ public final class GenerateInfoPlistProjectMapper: ProjectMapping {
     public func map(project: Project) throws -> (Project, [SideEffectDescriptor]) {
         logger.debug("Transforming project \(project.name): Synthesizing Info.plist")
 
-        let results = try project.targets.values
-            .reduce(into: (targets: [String: Target](), sideEffects: [SideEffectDescriptor]())) { results, target in
+        let results = try project.targets
+            .reduce(into: (targets: [Target](), sideEffects: [SideEffectDescriptor]())) { results, target in
                 let (updatedTarget, sideEffects) = try map(target: target, project: project)
-                results.targets[updatedTarget.name] = updatedTarget
+                results.targets.append(updatedTarget)
                 results.sideEffects.append(contentsOf: sideEffects)
             }
         var project = project