--- a/Sources/TuistGenerator/Mappers/ResourcesProjectMapper.swift
+++ b/Sources/TuistGenerator/Mappers/ResourcesProjectMapper.swift
@@ -18,11 +18,11 @@ public class ResourcesProjectMapper: ProjectMapping { // swiftlint:disable:this 
         logger.debug("Transforming project \(project.name): Generating bundles for libraries'")
 
         var sideEffects: [SideEffectDescriptor] = []
-        var targets: [String: Target] = [:]
+        var targets: [Target] = []
 
-        for target in project.targets.values {
+        for target in project.targets {
             let (mappedTargets, targetSideEffects) = try mapTarget(target, project: project)
-            mappedTargets.forEach { targets[$0.name] = $0 }
+            mappedTargets.forEach { targets.append($0) }
             sideEffects.append(contentsOf: targetSideEffects)
         }