… Sources/TuistGenerator/Mappers/AutogeneratedSchemesProjectMapper.swift
… Sources/TuistGenerator/Mappers/AutogeneratedWorkspaceSchemeWorkspaceMapper.swift
… Sources/TuistKit/Mappers/TreeShakePrunedTargetsGraphMapper.swift
… Sources/TuistKit/ProjectEditor/ProjectEditorMapper.swift
… Sources/TuistKit/Services/ProjectAutomation+ManifestMapper.swift
… Tests/TuistDependenciesTests/Mappers/ExternalProjectsPlatformNarrowerGraphMapperTests.swift
… Tests/TuistDependenciesTests/Mappers/PruneOrphanExternalTargetsGraphMapperTests.swift
… Tests/TuistGeneratorIntegrationTests/Generator/MultipleConfigurationsIntegrationTests.swift
… Tests/TuistGeneratorIntegrationTests/Generator/StableStructureIntegrationTests.swift
… Tests/TuistGeneratorIntegrationTests/Generator/TestModelGenerator.swift
… Tests/TuistGeneratorIntegrationTests/Generator/TuistGeneratorPerformanceTests.swift
… Tests/TuistGeneratorTests/Generator/SchemeDescriptorsGeneratorTests.swift
… Tests/TuistGeneratorTests/ProjectMappers/ExplicitDependencyGraphMapperTests.swift
… Tests/TuistGeneratorTests/ProjectMappers/GenerateInfoPlistProjectMapperTests.swift
… Tests/TuistGeneratorTests/ProjectMappers/ResourcesProjectMapperTests.swift
… Tests/TuistGeneratorTests/ProjectMappers/TargetActionDisableShowEnvVarsProjectMapperTests.swift
… Tests/TuistKitTests/Mappers/FocusTargetsGraphMappersTests.swift
… Tests/TuistKitTests/Mappers/Graph/TreeShakePrunedTargetsGraphMapperTests.swift
… Tests/TuistKitTests/ProjectEditor/ProjectEditorMapperTests.swift
… Tests/TuistLoaderTests/Loaders/ManifestModelConverterTests.swift