As noted [here](https://github.com/tuist/tuist/pull/6362), [this work](https://github.com/tuist/tuist/pull/6350) introduced a regression causing generated Xcode projects and workspaces to be non-deterministic across runs. This PR fixes the issue instead of reverting the work.

The issue happened because I changed `targets: [Target]` into `targets: [String: Target]` in `TuistGraph.Project`, causing the order to get lost. I reverted it to `[Target]`.

I also moved the integration tests at `TuistIntegrationTests` under `TuistGeneratorIntegrationTests` to ensure they run as part of every PR and that we don't introduce regressions in this area again.