@pepicrft 2024-06-03
I'm investigating if we can do something to prevent introducing performance regressions, which are quite bad. Here are some numbers of before and after:

#### 20 projects, 100 targets per project, 100 sources per target
- Before changes
  - Time (mean ± σ):     12.088 s ±  0.858 s    [User: 37.930 s, System: 3.431 s]
  - Range (min … max):   11.083 s … 13.085 s    5 runs
- After changes
  - Time (mean ± σ):     31.776 s ±  4.534 s    [User: 121.451 s, System: 5.234 s]
  - Range (min … max):   26.957 s … 38.294 s    5 runs